일단 단순한 지도 조건은 마무리 했고, 플레이어 위치, 색 정보까지 마무리함.

+++ src/include/struct.h
@@ -36,8 +36,8 @@ typedef struct s_map
 	unsigned int floor_color;
 	unsigned int cell_color;
 	
-	double	px;
-	double	py;
+	unsigned int px;
+	unsigned int py;
 }	t_map;
 
 typedef struct s_color
